update breadcrumbs to handle doubly nested routes
